Updated rocks
Works a little more now. Needs to become a pointer.
Big rocks now flies in random directions. Still spawns at Point(0,0).
I need to make medium rocks and small rocks work in a similar fashion. We should also have rocks be a pointer so we can update the size and type of the rock.