Update
asteroids is now working in it's glory.
A few things to change, rocks should technically follow the rules with the left rock spawning 3 spaces to the right and the right doing the same in the opposite direction.
I need to also add some extra non required "Stretch" code, to spice up the game.

Files have been further organized, comments and lines either added or removed for clarity! Cheers!